Fixes/workaround for CJS consumption of components with nhsuk-frontend JS references. Better align components and stories to their folders/categories.
