+++ .storybook/preview.ts
@@ -6,7 +6,15 @@ const preview: Preview = {
     actions: { argTypesRegex: '^on[A-Z].*' },
     options: {
       storySort: {
-        order: ['Welcome', 'Migration Guides', 'Form Elements', 'Content Presentation', 'Navigation', 'Components', 'Patterns'],
+        order: [
+          'Welcome',
+          'Migration Guides',
+          'Form Elements',
+          'Content Presentation',
+          'Navigation',
+          'Layout',
+          'Patterns',
+        ],
       },
     },
   },
