+++ src/components/content-presentation/hero/__tests__/Hero.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-import Hero from '../';
+import Hero from '..';
 
 describe('Hero', () => {
   it('matches snapshot', () => {
