+++ src/components/content-presentation/hero/index.ts
