+++ src/components/content-presentation/icons/BaseIcon.tsx
