+++ src/components/content-presentation/icons/__tests__/Icons.test.tsx
