+++ src/components/content-presentation/icons/index.ts
