+++ src/components/content-presentation/icons/individual/ArrowLeft.tsx
