+++ src/components/content-presentation/icons/individual/ArrowRight.tsx
