+++ src/components/content-presentation/icons/individual/ArrowRightCircle.tsx
