+++ src/components/content-presentation/icons/individual/ChevronRight.tsx
