+++ src/components/content-presentation/icons/individual/ChevronRightCircle.tsx
