+++ src/components/content-presentation/icons/individual/Close.tsx
