+++ src/components/content-presentation/icons/individual/Cross.tsx
