+++ src/components/content-presentation/icons/individual/Minus.tsx
