+++ src/components/content-presentation/icons/individual/Plus.tsx
