+++ src/components/content-presentation/icons/individual/SmallEmdash.tsx
