+++ src/components/content-presentation/icons/individual/Tick.tsx
