+++ src/components/navigation/action-link/ActionLink.tsx
@@ -1,6 +1,6 @@
 import React, { FC } from 'react';
 import classNames from 'classnames';
-import { ArrowRightCircle } from '@components/icons';
+import { ArrowRightCircle } from '@components/content-presentation/icons';
 import { AsElementLink } from '@util/types/LinkTypes';
 
 const ActionLink: FC<AsElementLink<HTMLAnchorElement>> = ({
