+++ src/components/navigation/header/components/NavDropdownMenu.tsx
@@ -1,7 +1,7 @@
 'use client';
 import React, { FC, HTMLProps, useContext, useEffect, MouseEvent } from 'react';
 import HeaderContext, { IHeaderContext } from '../HeaderContext';
-import { ChevronDown as ChevronDownIcon } from '@components/icons';
+import { ChevronDown as ChevronDownIcon } from '@components/content-presentation/icons';
 
 export interface NavDropdownMenuProps extends HTMLProps<HTMLButtonElement> {
   type?: 'button' | 'submit' | 'reset';
