+++ src/components/navigation/header/components/Search.tsx
@@ -1,6 +1,6 @@
 import React, { FC, HTMLProps, useContext, useEffect } from 'react';
 import classNames from 'classnames';
-import { Search as SearchIcon } from '@components/icons';
+import { Search as SearchIcon } from '@components/content-presentation/icons';
 import HeaderContext, { IHeaderContext } from '../HeaderContext';
 
 export interface SearchProps extends HTMLProps<HTMLInputElement> {
