+++ src/global.d.ts
@@ -1,5 +1,5 @@
-declare module 'nhsuk-frontend/packages/components/header/header.js';
-declare module 'nhsuk-frontend/packages/components/checkboxes/checkboxes.js';
-declare module 'nhsuk-frontend/packages/components/radios/radios.js';
-declare module 'nhsuk-frontend/packages/components/character-count/character-count.js';
-declare module 'nhsuk-frontend/packages/components/tabs/tabs.js';
+declare module 'nhsuk-frontend/packages/components/header/header';
+declare module 'nhsuk-frontend/packages/components/checkboxes/checkboxes';
+declare module 'nhsuk-frontend/packages/components/radios/radios';
+declare module 'nhsuk-frontend/packages/components/character-count/character-count';
+declare module 'nhsuk-frontend/packages/components/tabs/tabs';
